+++ NEWS.md
@@ -1,6 +1,10 @@
 # 0.1.0.9000
 
 * Fix bug where you can't rename objects via name in `gcs_upload`
+* Add `gcs_save` to store R session data in cloud
+* Add `gcs_load` to restore session data stored with `gcs_save`
+* Fix resetting of `options(googleAuthR.rawResponse = TRUE)` when using `gcs_get_object`
+* Add URLencoding to `object_name` in `gcs_get_object`
 
 # googleCloudStorageR 0.1.0